refactor: centralized search_path function
